// packages/vgrammar-plot/src/interface.ts
export type PlotAreaEncoderSpec = Omit<BasicEncoderSpecMap['area'], 'x' | 'y'> & {
  x?: number | number[];
  y?: number | number[];
};
// ...
export type WithDefaultEncode<T, K extends string> = {
  [Key in K]?: Key extends keyof T ? ISemanticEncodeValue<T[Key]> : ISemanticEncodeValue<string | number>;
};
export type ISemanticEncodeValue<T> = T extends any[]
  ? string[] | ((datum: any) => any)[]
  : string | ((datum: any) => T);
export type ISemanticEncodeSpec<T> = {
  [Key in keyof T]?: ISemanticEncodeValue<T[Key]>;
};
export type ISemanticStyle<T, K extends string> = Omit<T, K>;

export type SemanticTooltipOption = {
  disableGraphicTooltip?: boolean;
  disableDimensionTooltip?: boolean;
  staticTitle?: string;
  staticContentKey?: string | string[];
  title?: ISemanticEncodeValue<string | number>;
  content?: Array<{
    key?: ISemanticEncodeValue<string | number>;
    value?: ISemanticEncodeValue<string | number>;
    symbol?: ISemanticEncodeValue<string>;
  }>;
};
// ...
export type AreaEncodeChannels = 'x' | 'y' | 'color' | 'group';